requirements/base.txt
Django==1.8.1
Werkzeug==0.10.4
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==5.4.0
django-bootstrap3-datetimepicker==2.2.3
django-extensions==1.5.3
django-model-utils==2.3.1
django-pandas==0.3.0
django-suit==0.2.14
django-toolbelt==0.0.1
gnureadline==6.3.3
gunicorn==19.3.0
ipython==3.1.0
numpy==1.9.2
pandas==0.16.2
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
static3==0.6.1
wsgiref==0.1.2
